% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{methods}
\name{.assignQuickPlot}
\alias{.assignQuickPlot}
\alias{.assignQuickPlot,character,ANY-method}
\alias{.assignQuickPlot,character,missing-method}
\title{Assign to the internal \code{quickPlot} environment.}
\usage{
.assignQuickPlot(x, value, ...)

\S4method{.assignQuickPlot}{character,ANY}(x, value, ...)

\S4method{.assignQuickPlot}{character,missing}(x, value, ...)
}
\arguments{
\item{x}{a variable name, given as a character string.
No coercion is done, and the first element of a character vector
of length greater than one will be used, with a warning.}

\item{value}{The object to assign. If this is missing, values will be found
with \code{get(x)} in the same environment as the calling
environment.}

\item{...}{Additional arguments to pass to \code{assign}.}
}
\value{
Only used for its side effect, namely the object assigned to the \code{.quickPlotEnv}
}
\description{
Internal function. Simple wrapper for \code{\link{assign}}.
}
\author{
Alex Chubaty
}
\keyword{internal}
