% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_jitter_plot.R
\name{box_jitter_plot}
\alias{box_jitter_plot}
\title{Plot a boxplot with a jitterplot}
\usage{
box_jitter_plot(
  dataset,
  var_x,
  var_y,
  x_label,
  y_label,
  alpha_boxplot = 0.3,
  alpha_jitter = 0.3,
  width_jitter = 0.1
)
}
\arguments{
\item{dataset}{The dataset containing the variables for plotting in a tidy format. Names come from "dataset used to plot"}

\item{var_x}{Name of the column containing the categorical variable to be plot in x axis. Name comes from "variable x"}

\item{var_y}{Name of the column containing the numerical variable to be plot in y axi. Name comes from "variable y"}

\item{x_label, y_label}{String indicating the name of the labels for x and y axis. Names come from "label of axis x or y"}

\item{alpha_boxplot, alpha_jitter}{Transparency parameter of the boxplot and jitterplot. Default is set to 0.3. Names come from "alpha of boxplot or jitterplot"}

\item{width_jitter}{Width parameter of the jitterplot. Default is set to 0.1. Name comes from "width of jitterplot"}
}
\value{
A ggplot2 histogram + jitter plot of a numerical variable with the mean marked as a black asterisk.
}
\description{
This function creates a boxplot with a jitter plot to observe the distribution of a numeric variable
across one or more categorical variables. Furthermore, the mean is denoted as a black asterisk.
}
\examples{
#Generate a box + jitter plot with the mpg dataset. The resulting plot can be further modified
# with ggplot2 functions to improve the visualization
box_jitter_plot(ggplot2::mpg,
    var_x = model,
    var_y = displ,
    x_label = "Car model",
    y_label = "Engine displacement (liters)",
    alpha_boxplot = 0.4,
    alpha_jitter = 0.4,
    width_jitter = 0.15) +
ggplot2::coord_flip()

}
