% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{coordinates,ANY-method}
\alias{coordinates,ANY-method}
\title{Extract coordinates from a variety of spatial objectgs}
\usage{
\S4method{coordinates}{ANY}(obj, ...)
}
\arguments{
\item{obj}{An object from which to extract the coordinates (e.g., \code{sf}, \code{sp})}

\item{...}{Ignored.}
}
\value{
A 2 column matrix of coordinates (x and y)
}
\description{
This will extract using \code{terra::crds}, \code{sf::st_coordinates} and
\code{raster::coordinates}. Other packages can create methods, as this is
generic.
}
