% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{fortify}
\alias{fortify}
\alias{.fortify}
\title{Fortify - i.e,. convert an arbitrary object to a data.frame-like object}
\usage{
.fortify(x, matchFortify = TRUE, simple = FALSE,
  maxNumPolygons = getOption("quickPlot.maxNumPolygons", 3000))
}
\description{
This only deals with SpatialPolygons.
}
\keyword{internal}
