% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboxplot.R
\name{ggboxplot}
\alias{ggboxplot}
\title{Title}
\usage{
ggboxplot(
  data,
  categorical_variable,
  numeric_variable,
  xlab = categorical_variable,
  ylab = numeric_variable
)
}
\arguments{
\item{data}{a data frame or a tibble from which the input variables are taken}

\item{categorical_variable}{a string that contains the name of a categorical variable of interest that is used for the boxplot and is mapped to the x axis}

\item{numeric_variable}{a string that contains a numeric variable of interest that is used for the boxplot and is mapped to the y axis}

\item{xlab}{a string that contains the text for the x axis. If not specified, the default value is the name of the categorical variable}

\item{ylab}{a string that contains the text for the y axis. If not specified, the default value is the name of the numeric variable}
}
\value{
ggboxplot returns an object of class ggplot.
}
\description{
Title
}
