% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-classes.R
\docType{class}
\name{griddedClasses-class}
\alias{griddedClasses-class}
\alias{griddedClasses}
\title{The \code{griddedClasses} class}
\description{
This class is the union of several spatial objects from \pkg{raster} and
\pkg{sp} packages.
}
\details{
Members:

\itemize{
  \item \code{RasterLayer}, \code{RasterLayerSparse}, \code{RasterStack};
}

Notably missing is \code{RasterBrick}, for now.
}
\seealso{
\code{\link{quickPlotClasses}}
}
\author{
Eliot McIntire
}
