% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{.hasBbox}
\alias{.hasBbox}
\title{Test whether class has bbox method}
\usage{
.hasBbox(z, objClass, objName, objEnv)
}
\arguments{
\item{z}{Logical, whether this object is a SpatialObject}

\item{objClass}{The class of the object}

\item{objName}{The character string name of the object}

\item{objEnv}{The environment where the object can be found}
}
\description{
For internal use only.
}
