% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\docType{methods}
\name{.identifyGrobToPlot}
\alias{.identifyGrobToPlot}
\alias{.identifyGrobToPlot,.quickPlotGrob,list-method}
\title{Identify where to get the grob from}
\usage{
.identifyGrobToPlot(grobNamesi, toPlot, takeFromPlotObj)

\S4method{.identifyGrobToPlot}{.quickPlotGrob,list}(grobNamesi, toPlot,
  takeFromPlotObj)
}
\arguments{
\item{grobNamesi}{name of the object to plot}

\item{toPlot}{list containing the objects to plot, made as a call to the
\code{Plot} function}

\item{takeFromPlotObj}{Logical. Should the data come from the argument passed
into Plot (\code{TRUE}), or from the (\code{.quickPlotEnv})
(\code{FALSE}).}
}
\description{
Internal function.
}
\details{
Because the Plot function can use the global environment as a source of
objects to plot, not just the call itself, this function identifies where
the data for the grob should come from, the current call or the global
environment.
}
\author{
Eliot McIntire
}
\keyword{internal}
