% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-other.R
\name{isRstudioServer}
\alias{isRstudioServer}
\title{Determine if current session is RStudio Server}
\usage{
isRstudioServer()
}
\description{
Determine if current session is RStudio Server
}
\examples{
isRstudioServer() # returns FALSE or TRUE
}
