% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{.makeList}
\alias{.makeList}
\alias{.makeList,list-method}
\title{Convert \code{plotArgs} to list of lists}
\usage{
.makeList(plotArgs, numQuickPlotObjects)

\S4method{.makeList}{list}(plotArgs, numQuickPlotObjects)
}
\arguments{
\item{plotArgs}{list. Any arguments that the the grid package can accept for
the specific grob types, e.g., \code{rasterGrob}, \code{polygonGrob}, etc.}

\item{numQuickPlotObjects}{Numeric. The number of \code{.quickPlotObjects}.
This can't easily be deduced from the \code{plotArgs} because
of the \code{RasterStack}s. So passed manually.}
}
\description{
Internal function. Take the inputs as \code{plotArgs} to the Plot function, and make
them a list of length \code{numQuickPlotObjects} entries of lists.
}
\author{
Eliot McIntire
}
\keyword{internal}
