% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\docType{methods}
\name{.makeQuickPlot}
\alias{.makeQuickPlot}
\alias{.makeQuickPlot,list,list-method}
\alias{.makeQuickPlot,list,missing-method}
\alias{.makeQuickPlot,missing,missing-method}
\title{Make a \code{.quickPlot} class object}
\usage{
.makeQuickPlot(plotObjects, plotArgs, whichQuickPlottables, ...)

\S4method{.makeQuickPlot}{list,list}(plotObjects, plotArgs,
  whichQuickPlottables, ...)

\S4method{.makeQuickPlot}{list,missing}(plotObjects, plotArgs,
  whichQuickPlottables, ...)

\S4method{.makeQuickPlot}{missing,missing}(plotObjects, plotArgs,
  whichQuickPlottables, ...)
}
\arguments{
\item{plotObjects}{list. Any plot objects.}

\item{plotArgs}{list. Any arguments that the the grid package can accept for
the specific grob types, e.g., rasterGrob, polygonGrob, etc.}

\item{whichQuickPlottables}{Logical indicating which objects in the
\code{Plot} call can be plotted by \code{Plot}.}

\item{...}{additional arguments. Currently nothing.}
}
\value{
A \code{\link{.quickPlot}} object, which has 2 slots, one for the plot arrangement
(i.e., layout and dimensions) and onefor all of the \code{quickPlotGrobs}
(stored as a quickPlotGrobList of lists \code{.quickPlotGrob} objects).
}
\description{
Builds a \code{.quickPlot} object from a list of objects.
}
\author{
Eliot McIntire
}
\keyword{internal}
