% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{.parseArgs}
\alias{.parseArgs}
\title{Parse arguments and find environments}
\usage{
.parseArgs(y, e, eminus1)
}
\arguments{
\item{y}{A character representation of the arguments passed to a function, e.g., \code{Plot}.}

\item{e}{Environment in which the function (e.g., \code{Plot}) was called.}

\item{eminus1}{The parent environment of \code{e}.}
}
\value{
A list of length 2, with names \code{objs} and \code{envs} giving the
standardized representation (i.e., replacing \verb{[[]]} with \code{$} notation for objects) of objects
and their layers (if \code{RasterStack}s).
}
\description{
Internal function used within \code{objectNames}.
}
\author{
Eliot McIntire and Alex Chubaty
}
\keyword{internal}
