% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plot.R
\name{quick_plot}
\alias{quick_plot}
\title{Quickly make a boxplot or scatterplot}
\usage{
quick_plot(
  x,
  y,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  interaction_variable = NULL,
  interaction_lab = NULL
)
}
\arguments{
\item{x}{a vector with class \code{numeric} or \code{factor}.}

\item{y}{a vector with class \code{numeric}.}

\item{xlab}{character indicating the x-axis label. Default is NULL.}

\item{ylab}{character indicating the y-axis label. Default is NULL.}

\item{title}{character indicating the main title. Default is NULL.}

\item{interaction_variable}{a vector with class \code{factor}. Default is NULL.}

\item{interaction_lab}{character indicating the legend label/title for the interaction variable. Default is NULL.}
}
\value{
A ggplot object.
}
\description{
Quickly make a boxplot or scatterplot
}
\examples{
quickplot(x = factor(rbinom(100, 1, 0.4)), y = c(1:100), xlab = "Knows R Studio (0=no, 1=yes)", ylab = "Final Grade (\%)")
}
