% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPlot-package.R
\docType{package}
\name{quickPlot-package}
\alias{quickPlot}
\alias{quickPlot-package}
\title{The \code{quickPlot} package}
\description{
A high-level plotting system, built using 'grid' graphics, that is optimized
for speed and modularity. This has great utility for quick visualizations
when testing code, with the key benefit that visualizations are updated
independently of one another.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://quickplot.predictiveecology.org}
  \item \url{https://github.com/PredictiveEcology/quickPlot}
  \item Report bugs at \url{https://github.com/PredictiveEcology/quickPlot/issues}
}

}
\author{
\strong{Maintainer}: Eliot J B McIntire \email{eliot.mcintire@canada.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})

Authors:
\itemize{
  \item Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})
}

Other contributors:
\itemize{
  \item His Majesty the King in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
