% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{.quickPlotEnv}
\alias{.quickPlotEnv}
\title{The quickPlot environment}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.quickPlotEnv
}
\description{
Environment used internally to store internal package objects and methods.
}
\keyword{internal}
