% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-classes.R
\docType{class}
\name{.quickPlotObjects-class}
\alias{.quickPlotObjects-class}
\alias{.quickPlotObjects}
\title{The \code{.quickPlotObjects} class}
\description{
This class contains the union of spatialObjects and several other plot-type objects.
Currently, this includes \code{SpatialPoints*}, \code{SpatialPolygons*},
\code{SpatialLines*}, \code{RasterLayer}, \code{RasterStack}, and \code{ggplot} objects.
These are the object classes that the \code{\link{Plot}} function can handle.
}
\seealso{
\code{\link{quickPlotClasses}}
}
\author{
Eliot McIntire
}
