% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-classes.R
\docType{class}
\name{.quickPlottables-class}
\alias{.quickPlottables-class}
\alias{.quickPlottables}
\title{The \code{.quickPlottables} class}
\description{
This class is the union of all \code{.quickPlotObjects} (e.g., \verb{RasterLayer*},
\verb{SpatialPoints*}, \verb{SpatialPolygons*}, \code{ggplot}, \code{hist}, etc.)
and \code{\link[=.quickPlot]{.quickPlot()}} class objects.
This allows replotting of a \code{\link[=.quickPlot]{.quickPlot()}} object.
}
\seealso{
\code{\link[=quickPlotClasses]{quickPlotClasses()}}
}
\author{
Eliot McIntire
}
\keyword{internal}
