% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPlot-package.R
\docType{data}
\name{sample-maps}
\alias{sample-maps}
\title{Dummy maps included with \code{quickPlot}}
\format{raster}
\description{
All maps included here are randomly generated maps created using
\code{SpaDES.tools::gaussMap()}.
These are located within the \code{maps} folder of the package, and are used
in the vignettes.
Use \code{system.file("maps", package = "quickPlot")} to locate the \file{maps/}
directory on your system.
}
\details{
\itemize{
  \item \code{DEM.tif}: converted to a a small number of discrete levels
        (in 100m hypothetical units).
  \item \code{habitatQuality.tif}: made to look like a continuous habitat surface,
         rescaled to 0 to 1.
  \item \code{forestAge.tif}: rescaled to possible forest ages in a boreal forest setting.
  \item \code{forestCover.tif}: rescaled to possible forest cover in a boreal forest setting.
  \item \code{percentPine.tif}: rescaled to percentages.
}
}
\keyword{maps}
