% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_graph.R
\name{scatter_graph}
\alias{scatter_graph}
\title{Produce a scatter plot}
\usage{
scatter_graph(data, x, y)
}
\arguments{
\item{data}{Dataframe containing the variables you would like to produce a graph for. I named this parameter data as it relates to the name dataframe which is what it uses. This allowed for the function to be more simple and readable.}

\item{x}{Variable you would like to apply on the x-axis. I named this parameter x as it is responsible for the values on the x-axis. This allowed for the function to be more straightforward and understandable.}

\item{y}{Variable you would like to apply on the y-axis. I named this parameter y as it is responsible for the values on the y-axis. This allowed for the function to be more straightforward and understandable.}
}
\value{
This function returns a scatter plot with a linear regression line as a result of utilizing ggplot2.
}
\description{
This function produces a scatter plot with a linear regression based on two numeric variables
}
\details{
The graph that is produced by this function places the x parameter on the x-axis, and the y parameter on the y-axis. These variables are determined based on the dataframe that is defined as the data parameter. The graph that is produced will contain an alpha transparency of 0.3, and a linear regression line will also be placed on the graph. The limits of the graph will be expended to an origin of (0,0), and a classic theme has been applied.
}
\examples{
scatter_graph(mtcars, mpg, hp)
scatter_graph(palmerpenguins::penguins, bill_length_mm, flipper_length_mm)
scatter_graph(palmerpenguins::penguins, body_mass_g, flipper_length_mm)

}
