% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-classes.R
\docType{class}
\name{sfClasses-class}
\alias{sfClasses-class}
\alias{sfClasses}
\title{The \code{sfClasses} class}
\description{
This class is the union of sf classes (sf & sfc)
}
\details{
Members:

\itemize{
  \item \code{sf}, \code{sfc};
}

Notably missing is \code{RasterBrick}, for now.
}
\seealso{
\code{\link{quickPlotClasses}}
}
\author{
Eliot McIntire
}
