% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{.updateQuickPlot}
\alias{.updateQuickPlot}
\alias{.updateQuickPlot,.quickPlot,list-method}
\alias{.updateQuickPlot,.quickPlot,missing-method}
\title{Merge two quickPlot objects}
\usage{
.updateQuickPlot(newSP, curr, ...)

\S4method{.updateQuickPlot}{.quickPlot,list}(newSP, curr, ...)

\S4method{.updateQuickPlot}{.quickPlot,missing}(newSP, curr, ...)
}
\arguments{
\item{newSP}{The "new" \code{.quickPlot} object.
I.e., the new merges and overwrites into current.}

\item{curr}{The "current" \code{.quickPlot} object.
I.e., the one to be merged into.}

\item{...}{Additional arguments. Currently none implemented.}
}
\description{
Merges two \code{.quickPlot} objects
}
\author{
Eliot McIntire
}
\keyword{internal}
