% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{extent}
\alias{extent}
\alias{extent,ANY-method}
\title{Get extent of a variety of spatial objects}
\usage{
\S4method{extent}{ANY}(x, ...)
}
\arguments{
\item{x}{The spatial object from which to extract the extent.}

\item{...}{Not used.}
}
\value{
Returns a list of length 4 with elements \code{xmin}, \code{xmax}, \code{ymin}, and \code{ymax},
in that order.
}
\description{
This is a wrapper around \code{terra::ext}, \code{sf::st_bbox}, and
\code{raster::extent}.
}
