% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helpers.R
\name{fortify}
\alias{fortify}
\alias{.fortify}
\title{Fortify}
\usage{
.fortify(
  x,
  matchFortify = TRUE,
  simple = FALSE,
  maxNumPolygons = getOption("quickPlot.maxNumPolygons", 3000)
)
}
\description{
Convert an arbitrary object to a \code{data.frame}-like object.
}
\note{
This only deals with \code{SpatialPolygons} objects.
}
\keyword{internal}
