% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-classes.R
\docType{class}
\name{.quickPlottables-class}
\alias{.quickPlottables-class}
\alias{.quickPlottables}
\title{The \code{.quickPlottables} class}
\description{
This class is the union of all .quickPlotObjects (e.g., \code{RasterLayer*},
\code{SpatialPoints*}, \code{SpatialPolygons*}, \code{ggplot}, \code{hist}, etc.)
and \code{\link{.quickPlot}} class objects.
This allows replotting of a \code{\link{.quickPlot}} object.
}
\section{Slots}{

\describe{
\item{\code{members}}{\code{\link{.quickPlotObjects}} and \code{\link{.quickPlot}}}
}}

\seealso{
\code{\link{quickPlotClasses}}
}
\author{
Eliot McIntire
}
\keyword{internal}
